refactor: all static types in domain. simplify architecture
