+++ src/domain/runtime_types.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from dataclasses import dataclass, field
 from typing import Any
 
 import numpy as np
@@ -7,14 +6,13 @@ import numpy as np
 from domain.io_types import FrameData
 
 
-@dataclass
 class CctaRuntimeData:
     def __init__(self):
         self.metadata: dict = {}
         self.volume: np.ndarray | None = None  # (Z, Y, X) int16 HU
         self.voxel_spacing: tuple[float, float, float] | None = None  # (dz, dy, dx) mm
         self.mask: np.ndarray | None = None  # (Z, Y, X) uint8 label values
-        self.labels: list[int] = field(default_factory=list)  # non-background labels present in mask
+        self.labels: list[int] = []  # non-background labels present in mask
 
 
 class RuntimeData:
