+++ src/input_output/input/ccta_io.py
